services/network_connector/requirements.txt
paho-mqtt
redis
systemd-python